move segmentaion metrics to performance_metrics
